Revert "Feat/info rx"
