+++ MateList.xcodeproj/xcshareddata/xcschemes/MateList.xcscheme
@@ -28,17 +28,6 @@
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
-         <TestableReference
-            skipped = "NO"
-            parallelizable = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "0458E4012960A9C4002C1A4A"
-               BuildableName = "MateListTests.xctest"
-               BlueprintName = "MateListTests"
-               ReferencedContainer = "container:MateList.xcodeproj">
-            </BuildableReference>
-         </TestableReference>
       </Testables>
    </TestAction>
    <LaunchAction
