+++ MateList/Data/Repository/Common/IDFirestoreRepository.swift
@@ -11,106 +11,6 @@ import FirebaseAuth
 
 
 class IDFirestoreRepository: IDFirestore {
-    // MARK: Check
-    func isExistUser(uid: String) -> Observable<Bool> {
-        return Observable.create { observer in
-            FireStoreService.db.collection("User").whereField("uid", isEqualTo: uid).getDocuments { (querySnapshot, err) in
-                guard let documents = querySnapshot?.documents else { return }
-                let data = documents.map{$0["uid"]!}
-                
-                observer.onNext(!data.isEmpty)
-            }
-            return Disposables.create()
-        }
-    }
-    
-    func isBlackUser(uid: String) -> Observable<Bool> {
-        return Observable.create { observer in
-            FireStoreService.db.collection("BlackList").whereField(uid, isEqualTo: true).getDocuments() { (querySnapshot, err) in
-                if let err = err {
-                    print("Error getting documents: \(err)")
-                } else {
-                    for _ in querySnapshot!.documents {
-                        observer.onNext(true)
-                    }
-                }
-            }
-            return Disposables.create()
-        }
-    }
-    
-    // MARK: Delete
-    func deleteAllPost(uid: String) {
-        var userPostsID: [String] = []
-        FireStoreService.db.collection("Post").whereField("uid", isEqualTo: uid).getDocuments() {(querySnapshot, err) in
-            if let err = err {
-                print("Error getting documents: \(err)")
-            } else {
-                for document in querySnapshot!.documents {
-                    print("\(document.documentID) => \(document.data())")
-                    userPostsID.append(document.documentID)
-                }
-            }
-        }
-        
-        for document_id in userPostsID {
-            FireStoreService.db.collection("Post").document(document_id).delete() { err in
-                if let err = err {
-                    print("Error removing document: \(err)")
-                } else {
-                    print("Document successfully removed!")
-                }
-            }
-        }
-    }
-    
-    func deleteAllScrap(uid: String) {
-        FireStoreService.db.collection("User").document(uid).collection("Scrap").getDocuments() {(querySnapshot, err) in
-            if let err = err {
-                print("Error getting documents: \(err)")
-            } else {
-                for document in querySnapshot!.documents {
-                    FireStoreService.db.collection("User").document(uid).collection("Scrap").document(document.documentID).delete() { err in
-                        if let err = err {
-                            print("Error removing document: \(err)")
-                        } else {
-                            print("Document successfully removed!")
-                        }
-                    }
-                }
-            }
-        }
-    }
-    
-    func deleteAllHabit(uid: String) {
-        FireStoreService.db.collection("User").document(uid).collection("HabitCheck").getDocuments() {(querySnapshot, err) in
-            if let err = err {
-                print("Error getting documents: \(err)")
-            } else {
-                for document in querySnapshot!.documents {
-                    FireStoreService.db.collection("User").document(uid).collection("HabitCheck").document(document.documentID).delete() { err in
-                        if let err = err {
-                            print("Error removing document: \(err)")
-                        } else {
-                            print("Document successfully removed!")
-                        }
-                    }
-                }
-            }
-        }
-    }
-    
-     func deleteUser(uid: String) {
-        FireStoreService.db.collection("User").document(uid).delete() { err in
-            if let err = err {
-                print("Error removing document: \(err)")
-            } else {
-                print("Document successfully removed!")
-            }
-        }
-    }
-    
-    // MARK: Read
     static func userName(uid: String) -> Observable<String> {
         return Observable.create { observer in
             let docRef = FireStoreService.db.collection("User")
