--- MateList/Data/Repository/Setting/Protocol/SettingRepositoryProtocol.swift
@@ -1,16 +1,0 @@
-//
-//  SettingRepositoryProtocol.swift
-//  MateList
-//
-//  Created by 강민규 on 2023/01/02.
-//
-
-import FirebaseAuth
-
-protocol SettingRepositoryProtocol {
-    //func googleCredential() -> AuthCredential
-    //func appleCredential() -> AuthCredential
-    func authSignIn(credential: AuthCredential)
-    func authSignOut()
-    func registUser(user: User)
-}
