+++ MateList/Domain/Repository+Interface/SampleRepositoryInterface.swift
@@ -6,4 +6,3 @@
 //
 
 import Foundation
-
