--- MateList/Presentation/Setting/Coordinator/SampleCoordinator.swift
@@ -1,8 +1,0 @@
-//
-//  SampleCoordinator.swift
-//  MateList
-//
-//  Created by 강민규 on 2022/12/25.
-//
-
-import Foundation
