--- MateList/Presentation/Setting/View/NickNameView.swift
@@ -1,18 +1,0 @@
-//
-//  NickNameView.swift
-//  MateList
-//
-//  Created by 강민규 on 2022/12/27.
-//
-
-import UIKit
-
-final class NickNameView: BaseView {
-    override func configureUI() {
-        self.backgroundColor = .brown
-    }
-    
-    override func bind() {
-        print()
-    }
-}
