--- MateList/Presentation/Setting/ViewModel/OtherButton/SurveyButtonViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  SurveyButtonViewModel.swift
-//  MateList
-//
-//  Created by 강민규 on 2023/01/04.
-//
-
-import Foundation
